+++ packages/ui/src/settings/ExpertTab.test.tsx
@@ -50,6 +50,8 @@ function render(state: ExpertState | undefined): void {
         onRecheck={() => {}}
         onAssess={() => {}}
         onClearAssessment={() => {}}
+        onMeasureCost={() => {}}
+        onClearPricing={() => {}}
       />,
     ),
   )